fix: load model in float16 for Render to avoid OOM on 512MB free tier

+++ build_model_cache.py
@@ -4,8 +4,9 @@ from transformers import VisionEncoderDecoderModel, ViTImageProcessor, AutoToken
 
 print("--- DOWNLOADING MODEL FOR CACHE SPEEDUPS ---")
 try:
+    import torch
     # Pre-cache vit-gpt2 for Render execution speed
-    VisionEncoderDecoderModel.from_pretrained("nlpconnect/vit-gpt2-image-captioning")
+    VisionEncoderDecoderModel.from_pretrained("nlpconnect/vit-gpt2-image-captioning", torch_dtype=torch.float16)
     ViTImageProcessor.from_pretrained("nlpconnect/vit-gpt2-image-captioning")
     AutoTokenizer.from_pretrained("nlpconnect/vit-gpt2-image-captioning")
     print("--- CACHING SUCCESSFUL ---")
